Vagrant fixes after testing

+++ vagrant/deploy.sh
@@ -23,16 +23,19 @@ export CONCOURSE_WEB_USER CONCOURSE_WEB_PASSWORD
 export VAGRANT_DEFAULT_PROVIDER="aws"
 export VAGRANT_BOX_NAME="aws_vagrant_box"
 
-if aws ec2 describe-key-pairs --key-name "${VAGRANT_SSH_KEY_NAME}" >/dev/null 2>&1 ; then
-  # Remove the aws key pair
-  aws ec2 delete-key-pair --key-name "${VAGRANT_SSH_KEY_NAME}"
-fi
- 
-# Create the key pair online.
-aws ec2 create-key-pair --key-name "${VAGRANT_SSH_KEY_NAME}" | jq -r ".KeyMaterial" > "${VAGRANT_SSH_KEY}"
+if ! vagrant status 2>&1 | grep "running\ (aws)" >/dev/null; then
+  echo "Concourse Bootstrap VM is not running."
+  if aws ec2 describe-key-pairs --key-name "${VAGRANT_SSH_KEY_NAME}" >/dev/null 2>&1 ; then
+    echo "Cleaning up old key pair..."
+    aws ec2 delete-key-pair --key-name "${VAGRANT_SSH_KEY_NAME}"
+  fi
+  echo "Creating new key pair..."
+  # Create the key pair online.
+  aws ec2 create-key-pair --key-name "${VAGRANT_SSH_KEY_NAME}" | jq -r ".KeyMaterial" > "${VAGRANT_SSH_KEY}"
 
-# Secure the local key.
-chmod 600 "${VAGRANT_SSH_KEY}"
+  # Secure the local key.
+  chmod 600 "${VAGRANT_SSH_KEY}"
+fi
 
 # Install aws dummy box if not present
 if ! vagrant box list | grep -qe "^${VAGRANT_BOX_NAME} "; then
@@ -42,10 +45,12 @@ fi
 
 vagrant up
 
-echo "Setting up SSH tunnel to concourse..."
-
-vagrant ssh -- -L 8080:127.0.0.1:8080 -fN
-
+if ! pgrep -f "ssh.*${VAGRANT_SSH_KEY_NAME}.*-L 8080:127.0.0.1:8080 -fN" >/dev/null; then
+  echo "Setting up SSH tunnel to concourse..."
+  vagrant ssh -- -L 8080:127.0.0.1:8080 -fN
+else
+  echo "SSH tunnel to concourse already running."
+fi
 timeout=180
 deadline=$(($(date +%s) + timeout))
 echo -n "Waiting for concourse to start for ${timeout}s..."
